+++ framework/validators/Validator.php
@@ -80,7 +80,7 @@ class Validator extends Component
         ],
         'unique' => 'yii\validators\UniqueValidator',
         'url' => 'yii\validators\UrlValidator',
-        'ip' => 'yii\validators\IpAddressValidator',
+        'ip' => 'yii\validators\IpValidator',
     ];
     /**
      * @var array|string attributes to be validated by this validator. For multiple attributes,